Moved config for navigation tracking
