+++ posthog-react-native/index.ts
@@ -8,3 +8,4 @@ export * from './src/hooks/useFeatureFlags'
 export * from './src/hooks/useFeatureFlag'
 export * from './src/hooks/usePostHog'
 export * from './src/PostHogProvider'
+export * from './src/types'
